Move oauth handler to controller
